#  :wrench: improving typechecker and lint local performance