--- a/docker-compose.yml
+++ b/docker-compose.yml
@@ -15,4 +15,3 @@ services:
     volumes:
       - '.:/app'
     restart: 'no'
-