--- a/package.json
+++ b/package.json
@@ -44,10 +44,10 @@
     "vrt:docker": "./bin/run-vrt",
     "rebuild-electron": "./node_modules/.bin/electron-rebuild -f -m ./packages/loot-core",
     "rebuild-node": "yarn workspace loot-core rebuild",
-    "lint": "eslint . --max-warnings 0",
-    "lint:verbose": "DEBUG=eslint:cli-engine eslint . --max-warnings 0",
+    "lint": "prettier --check . && eslint . --max-warnings 0",
+    "lint:fix": "prettier --check --write . && eslint . --max-warnings 0 --fix",
     "install:server": "yarn workspaces focus @actual-app/sync-server --production",
-    "typecheck": "yarn tsc && tsc-strict",
+    "typecheck": "yarn tsc --incremental && tsc-strict",
     "jq": "./node_modules/node-jq/bin/jq",
     "prepare": "husky"
   },
@@ -60,7 +60,6 @@
     "eslint-import-resolver-typescript": "^4.2.2",
     "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-jsx-a11y": "^6.10.2",
-    "eslint-plugin-prettier": "5.2.3",
     "eslint-plugin-react": "^7.37.4",
     "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-rulesdir": "^0.2.2",
@@ -84,7 +83,7 @@
     "node": ">=18.0.0"
   },
   "lint-staged": {
-    "*.{js,jsx,ts,tsx,md,json}": "prettier --write"
+    "*.{js,jsx,ts,tsx,md,json,yml}": "prettier --write"
   },
   "packageManager": "yarn@4.7.0",
   "browserslist": [