--- a/packages/desktop-client/bin/remove-untranslated-languages
+++ b/packages/desktop-client/bin/remove-untranslated-languages
@@ -20,7 +20,7 @@ const processTranslations = () => {
 
     console.log(`en.json has ${enKeysCount} keys.`);
 
-    files.forEach((file) => {
+    files.forEach(file => {
       if (file === 'en.json' || path.extname(file) !== '.json') return;
 
       if (file.startsWith('en-')) {
@@ -34,7 +34,9 @@ const processTranslations = () => {
 
       // Calculate the percentage of keys present compared to en.json
       const percentage = (fileKeysCount / enKeysCount) * 100;
-      console.log(`${file} has ${fileKeysCount} keys (${percentage.toFixed(2)}%).`);
+      console.log(
+        `${file} has ${fileKeysCount} keys (${percentage.toFixed(2)}%).`,
+      );
 
       if (percentage < 50) {
         fs.unlinkSync(filePath);