--- a/packages/desktop-client/globals.d.ts
+++ b/packages/desktop-client/globals.d.ts
@@ -1,4 +1,3 @@
-// eslint-disable-next-line import/no-unresolved
 import { type CSSObject } from '@emotion/css/dist/declarations/src/create-instance';
 
 // Allow images to be imported