--- a/packages/desktop-client/src/fonts.scss
+++ b/packages/desktop-client/src/fonts.scss
@@ -3,4 +3,4 @@
 );
 @include variable.default;
 
-@import "@fontsource/redacted-script";
+@import '@fontsource/redacted-script';