--- a/packages/desktop-electron/.babelrc
+++ b/packages/desktop-electron/.babelrc
@@ -1,5 +1,3 @@
 {
-    plugins: [
-      'babel-plugin-transform-es2015-modules-commonjs'
-    ]
+  "plugins": ["babel-plugin-transform-es2015-modules-commonjs"]
 }