--- a/packages/desktop-electron/loading.html
+++ b/packages/desktop-electron/loading.html
@@ -1,4 +1,4 @@
-<!DOCTYPE html>
+<!doctype html>
 
 <body>
   <style type="text/css">
@@ -12,7 +12,5 @@
       color: #909090;
     }
   </style>
-  <div>
-  Waiting for development server...
-  </div>
+  <div>Waiting for development server...</div>
 </body>