--- a/packages/loot-core/src/mocks/files/default-budget-template/metadata.json
+++ b/packages/loot-core/src/mocks/files/default-budget-template/metadata.json
@@ -1,1 +1,6 @@
-{"id":"Default-Test-Db","budgetName":"Default Test Db","budgetVersion":"0.0.1","isCached":true}
+{
+  "id": "Default-Test-Db",
+  "budgetName": "Default Test Db",
+  "budgetVersion": "0.0.1",
+  "isCached": true
+}