--- a/packages/loot-core/tsconfig.api.json
+++ b/packages/loot-core/tsconfig.api.json
@@ -1,12 +1,17 @@
 {
-    "extends": "../../tsconfig.json",
-    "compilerOptions": {
-      "rootDir": "./src",
-      "declaration": true,
-      "emitDeclarationOnly": true,
-      "allowJs": false,
-      "noEmit": false,
-    },
-    "include": ["./typings", "./src/server/*"],
-    "exclude": ["**/node_modules/*", "**/build/*", "**/lib-dist/*", "./src/server/bench.ts"],
-  }
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "rootDir": "./src",
+    "declaration": true,
+    "emitDeclarationOnly": true,
+    "allowJs": false,
+    "noEmit": false
+  },
+  "include": ["./typings", "./src/server/*"],
+  "exclude": [
+    "**/node_modules/*",
+    "**/build/*",
+    "**/lib-dist/*",
+    "./src/server/bench.ts"
+  ]
+}