--- a/tsconfig.json
+++ b/tsconfig.json
@@ -42,6 +42,7 @@
   },
   "include": ["packages/**/*", "bin/*.ts"],
   "exclude": [
+    "**/.*/",
     "node_modules",
     "**/node_modules/*",
     "**/build/*",