--- a/upcoming-release-notes/4710.md
+++ b/upcoming-release-notes/4710.md
@@ -1,0 +1,6 @@
+---
+category: Maintenance
+authors: [MatissJanis]
+---
+
+improving typechecker and lint local performance