Applying some small quirks we can do to make lint & typechecker faster.

https://typescript-eslint.io/troubleshooting/typed-linting/performance/

---

## yarn typecheck
```
before:
38.29s user 2.34s system 169% cpu 24.005 total

with --incremental:
38.89s user 2.43s system 164% cpu 25.111 tota

with --incremental 2nd run:
17.20s user 1.74s system 155% cpu 12.193 total
```

### yarn lint
```
before:
110.63s user 7.36s system 205% cpu 57.361 total

after:
97.88s user 6.99s system 186% cpu 56.181 total
```

(not a massive difference here yet.. however, I have some more changes I'd like to do that should boost this quite a bit.. but those changes might be controversial, so I'll split them in a separate PR)